Implemented SortCommand Logic which now correctly sorts the list . Modified AddressBook.java to include a sortPersons method, which sorts it's persons list . Modified ModelManager.java to call sortPersons on it's addressBook . Modified UniquePersonList.java to implement the sortPersons method . Modified AddCommandTest to include the sortFilteredPersons method in it's stubModel
