+++ src/main/java/seedu/address/model/ModelManager.java
@@ -31,7 +31,7 @@ public class ModelManager extends ComponentManager implements Model {
     private static final Logger logger = LogsCenter.getLogger(ModelManager.class);
 
     private final AddressBook addressBook;
-    private final FilteredList<ReadOnlyPerson> filteredPersons;
+    private FilteredList<ReadOnlyPerson> filteredPersons;
 
     /**
      * Initializes a ModelManager with the given addressBook and userPrefs.
@@ -114,25 +114,9 @@ public class ModelManager extends ComponentManager implements Model {
 
     @Override
     public void sortFilteredPersonList() {
-
-        Comparator<ReadOnlyPerson> personComparator
-                = new Comparator<ReadOnlyPerson>() {
-
-            public int compare(ReadOnlyPerson person1, ReadOnlyPerson person2) {
-
-                String personName1 = person1.getName().toString();
-                String personName2 = person2.getName().toString();
-
-                //ascending order
-                return personName1.compareTo(personName2);
-
-                //descending order
-                //return personName2.compareTo(personName1);
-            }
-
-        };
-        
-        FXCollections.sort(filteredPersons, personComparator);
+        this.addressBook.sortPersons();
+        ObservableList<ReadOnlyPerson> sortedList = this.addressBook.getPersonList();
+        filteredPersons = new FilteredList<>(sortedList);
     }
 
     @Override
