+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -136,6 +136,11 @@ public class AddCommandTest {
         }
 
         @Override
+        public void sortFilteredPersonList() {
+            fail("This method should not be called.");
+        }
+
+        @Override
         public void deleteTag(Tag tag) {
             fail("This method should not be called.");
         }
